requirements.txt
python-docx
pyPDF2
python-pptx
pdfplumber
fitz